==> Course/01-intro-to-seminar.Rmd <==
# Introduction to Seminar {#intro-sem} 

In this course, students will learn the basics of data analysis using the R programming language. At the end of the course, students will be able to write an empirical seminar paper or BA-thesis using quantitative statistical modeling techniques. To achieve this, we will not only focus on how to apply this techniques in R, but also why certain approaches are chosen for certain problems and how to use them correctly with the aim of producing reliable statistical results.

The course starts with an introduction to exploratory data analysis; getting to know your data, your variables and the relationships between them. After this we will we go into statistical modeling. Before we can even start to model, we have to understand what modeling is, what approaches do exist and what we should and should not include in our model. You will learn how to use acyclical directed graphs (DAGs) to construct a model based on theoretical assumptions. We continue with a thorough introduction to simple and multiple linear regression. This will be the basis for more advanced topics that conclude the course, including introductions to logistic regression, prediction and machine learning.

## Prerequisites

Prior knowledge in the basics of using R is required. You should already know how the R syntax works, how to import, clean and manage data, how to compute descriptive statistics and how to create plots. Experience with the `tidyverse` packages is also required.
We highly encourage you to go through the following introduction to R written by Prof. Dr. Jasper Tjaden. This will equip you with all R knowledge you will need to successfully complete the seminar.
[Intro to R for Social Scientists](https://jaspertjaden.github.io/course-intro2r/)

Prepare by installing R and Rstudio. Here is a [link](https://rstudio-education.github.io/hopr/starting.html) that guides you through the installation process. If you already have an installation, you should make sure that it is updated to the most current versions.

You can also, this is optional, create a new R project in the folder which you will use for this seminar.


```{r load, include=FALSE, message=FALSE, warning=FALSE}
# Define the required packages
required_packages <- c("ggplot2", "dplyr", "tidyr", "dagitty", "mlbench", "caret", "broom", "performance", "ggeffects", "margins", "skimr", "tidyverse", "readxl", "lubridate", "gtsummary", "DataExplorer", "ggpubr", "knitr", "ggfortify")

# Check if all required packages are loaded
all(required_packages %in% search())

# Load any missing required packages
lapply(required_packages, require, character.only = TRUE)
```
